+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2026-03-15
+
+### Added
+- Standalone `Client` class that includes all Runner modules and builds `Elastic::AppSearch::Client` from stored `@opts`
+- Client specs covering initialization, default options, and runner method availability
+
 ## [0.1.1] - 2026-03-13
 
 ### Added